+++ crates/core/src/tool.rs
@@ -251,8 +251,11 @@ impl Tool {
     }
 
     /// Create an initial lock record.
-    pub fn create_lock_record(&self) -> LockRecord {
-        LockRecord::new(self.context.backend.clone(), self.proto.os, self.proto.arch)
+    pub fn create_locked_record(&self) -> LockRecord {
+        let mut record =
+            LockRecord::new(self.context.backend.clone(), self.proto.os, self.proto.arch);
+        record.inherit_options(&self.metadata.lock_options);
+        record
     }
 
     /// Register the tool by loading initial metadata and persisting it.
@@ -286,9 +289,9 @@ impl Tool {
         let mut inventory = self
             .proto
             .store
-            .create_inventory(self.get_id(), &metadata.inventory)?;
+            .create_inventory(self.get_id(), &metadata.inventory_options)?;
 
-        if let Some(override_dir) = &metadata.inventory.override_dir {
+        if let Some(override_dir) = &metadata.inventory_options.override_dir {
             let override_dir_path = override_dir.real_path();
 
             debug!(