+++ CHANGELOG.md
@@ -34,6 +34,8 @@
   - Added a `--type` option to `proto plugin add` and `proto plugin remove`.
 - Updated lockfiles to track `os` and `arch` information.
   - For existing entries, we attempt to backfill these values. If you run into issues, delete the lockfile and regenerate it.
+- **WASM API**
+  - Added a `RegisterToolOutput.lock_options` field, so that tools can customize how to interact with the lockfile.
 
 #### ⚙️ Internal